Added triggerChatRollFx
