+++ module/data/chat-message/actorRoll.mjs
@@ -1,3 +1,5 @@
+import { triggerChatRollFx } from '../../helpers/utils.mjs';
+
 const fields = foundry.data.fields;
 
 const targetsField = () =>
@@ -154,12 +156,7 @@ export default class DHActorRoll extends foundry.abstract.TypeDataModel {
             partData.roll = rerolled.toJSON();
         }
 
-        if (game.modules.get('dice-so-nice')?.active) {
-            const rerollPromises = rerolls.map(roll => game.dice3d.showForRoll(roll, game.user, true));
-            await Promise.allSettled(rerollPromises);
-        } else {
-            foundry.audio.AudioHelper.play({ src: CONFIG.sounds.dice });
-        }
+        await triggerChatRollFx(rerolls);
 
         return update;
     }
