+++ module/data/fields/action/damageField.mjs
@@ -72,9 +72,6 @@ export default class DamageField extends fields.SchemaField {
         damageConfig.source.message = messageId;
         damageConfig.directDamage = !!damageConfig.source?.message;
 
-        // if(damageConfig.source?.message && game.modules.get('dice-so-nice')?.active)
-        //     await game.dice3d.waitFor3DAnimationByMessageID(damageConfig.source.message);
-
         const damageResult = await CONFIG.Dice.daggerheart.DamageRoll.build(damageConfig);
         if (!damageResult) return false;
         if (damageResult.actionChatMessageHandled) config.actionChatMessageHandled = true;
